+++ internal/host/hostutil/host_utils_test.go
@@ -88,126 +88,6 @@ var _ = Describe("Installation Disk selection", func() {
 	}
 })
 
-var _ = Describe("host IP address families", func() {
-	It("host doesn't have interfaces", func() {
-		host := &models.Host{
-			Inventory: "{}",
-		}
-		v4, v6, err := GetAddressFamilies(host)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(v4).To(BeFalse())
-		Expect(v6).To(BeFalse())
-	})
-	It("error parsing inventory", func() {
-		host := &models.Host{
-			Inventory: "",
-		}
-		_, _, err := GetAddressFamilies(host)
-		Expect(err).Should(HaveOccurred())
-	})
-	It("host has only IPv4 addresses", func() {
-		host := &models.Host{
-			Inventory: `{
-				"interfaces":[
-					{
-						"ipv6_addresses":[],
-						"ipv4_addresses":[
-							"192.186.10.12/24"
-						]
-					}
-				]
-			}`,
-		}
-		v4, v6, err := GetAddressFamilies(host)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(v4).To(BeTrue())
-		Expect(v6).To(BeFalse())
-	})
-	It("host has only IPv6 addresses", func() {
-		host := &models.Host{
-			Inventory: `{
-				"interfaces":
-				[
-					{
-						"ipv6_addresses":[
-							"2002:db8::2/64"
-						],
-						"ipv4_addresses":[]
-					}
-				]
-			}`,
-		}
-		v4, v6, err := GetAddressFamilies(host)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(v4).To(BeFalse())
-		Expect(v6).To(BeTrue())
-	})
-	It("host has both IPv4 and IPv6 addresses on same interface", func() {
-		host := &models.Host{
-			Inventory: `{"interfaces":
-				[
-					{
-						"ipv4_addresses":[
-							"192.186.10.12/24"
-						],
-						"ipv6_addresses":[
-							"2002:db8::1/64"
-						]
-					}
-				]
-			}`,
-		}
-		v4, v6, err := GetAddressFamilies(host)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(v4).To(BeTrue())
-		Expect(v6).To(BeTrue())
-	})
-	It("host has both IPv4 and IPv6 addresses on different interfaces", func() {
-		host := &models.Host{
-			Inventory: `{
-				"interfaces":[
-					{
-						"ipv4_addresses":[
-							"192.186.10.12/24"
-						]
-					},
-					{
-						"ipv6_addresses":[
-							"2002:db8::1/64"
-						]
-					}
-				]
-			}`,
-		}
-		v4, v6, err := GetAddressFamilies(host)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(v4).To(BeTrue())
-		Expect(v6).To(BeTrue())
-	})
-	It("host has both IPv4 and IPv6 addresses on different interfaces, reverse order", func() {
-		host := &models.Host{
-			Inventory: `{
-				"interfaces":[
-					{
-						"ipv6_addresses":[
-							"2002:db8::1/64"
-						]
-					},
-					{
-						"ipv4_addresses":[
-							"192.186.10.12/24"
-						]
-					}
-				]
-			}`,
-		}
-		v4, v6, err := GetAddressFamilies(host)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(v4).To(BeTrue())
-		Expect(v6).To(BeTrue())
-	})
-})
-
 func TestHostUtil(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "HostUtil Tests")
