+++ internal/network/utils.go
@@ -1,43 +1,43 @@
 package network
 
 import (
-	"encoding/json"
-	"net/http"
-
-	"github.com/openshift/assisted-service/internal/common"
+	"github.com/openshift/assisted-service/internal/host/hostutil"
 	"github.com/openshift/assisted-service/models"
-	"github.com/sirupsen/logrus"
 )
 
-func IsIpv6OnlyHost(host *models.Host, log logrus.FieldLogger) (bool, error) {
-	if host.Inventory == "" {
-		return false, nil
+// GetHostAddressFamilies tests if a host has addresses in IPv4, in IPv6 family, or both
+func GetHostAddressFamilies(host *models.Host) (bool, bool, error) {
+	inventory, err := hostutil.UnmarshalInventory(host.Inventory)
+	if err != nil {
+		return false, false, err
 	}
-	var inventory models.Inventory
-	if err := json.Unmarshal([]byte(host.Inventory), &inventory); err != nil {
-		log.WithError(err).Warn("Can't unmarshal inventory")
-		return false, common.NewApiError(http.StatusBadRequest, err)
+	v4 := false
+	v6 := false
+	for _, i := range inventory.Interfaces {
+		v4 = v4 || len(i.IPV4Addresses) > 0
+		v6 = v6 || len(i.IPV6Addresses) > 0
+		if v4 && v6 {
+			break
+		}
 	}
-	var hasIpv4, hasIpv6 bool
-	for _, intf := range inventory.Interfaces {
-		hasIpv4 = hasIpv4 || len(intf.IPV4Addresses) > 0
-		hasIpv6 = hasIpv6 || len(intf.IPV6Addresses) > 0
-	}
-	return hasIpv6 && !hasIpv4, nil
+	return v4, v6, nil
 }
 
-func AreIpv6OnlyHosts(hosts []*models.Host, log logrus.FieldLogger) (bool, error) {
+// GetClusterAddressStack tests if all the hosts in a cluster have addresses in IPv4, in IPv6 family, or both (dual stack).
+// A dual-stack cluster requires all its hosts to be dual-stack.
+func GetClusterAddressStack(hosts []*models.Host) (bool, bool, error) {
 	if len(hosts) == 0 {
-		return false, nil
+		return false, false, nil
 	}
+	v4 := true
+	v6 := true
 	for _, h := range hosts {
-		ipv6Only, err := IsIpv6OnlyHost(h, log)
+		hostV4, hostV6, err := GetHostAddressFamilies(h)
 		if err != nil {
-			return false, err
+			return false, false, err
 		}
-		if !ipv6Only {
-			return false, nil
-		}
+		v4 = v4 && hostV4
+		v6 = v6 && hostV6
 	}
-	return true, nil
+	return v4, v6, nil
 }
